feat(components): split components into modules
